refactor(68): transpose `/api/auth/{provider}/{authorize|callback}` to `/api/auth/{authorize|callback}/{provider}`
