+++ back/src/auth/mod.rs
@@ -1,12 +1,8 @@
 use std::{error::Error as StdError, sync::OnceLock};
 
 use actix_web::{
-    dev::Payload,
-    error::ErrorInternalServerError,
-    http::header,
-    post,
-    web::{self, ServiceConfig},
-    Error as ActixError, FromRequest, HttpRequest, HttpResponse, Result as ActixResult,
+    dev::Payload, error::ErrorInternalServerError, http::header, web, Error as ActixError,
+    FromRequest, HttpRequest, Result as ActixResult,
 };
 use diesel::{Queryable, Selectable};
 use futures_util::future::LocalBoxFuture;
@@ -19,125 +15,20 @@ use utoipa::ToSchema;
 use uuid::Uuid;
 
 use crate::{
-    api::{json_payload_error_example, path_error_example},
     auth::{
         error::AuthErrorKind,
         oauth2_client::{ClientType, DynOAuth2Client},
     },
     config::AppConfiguration,
-    database::{self, Database, DatabaseError},
 };
 
 pub mod error;
 pub mod google;
 mod oauth2_client;
+pub mod routes;
 
 pub use oauth2_client::OAuth2Clients;
-
-/// Exposes all HTTP routes of this module.
-pub fn routes(clients: web::Data<OAuth2Clients>, service_cfg: &mut ServiceConfig) {
-    service_cfg.service(refresh_token);
-    if let Some(google) = clients.google() {
-        google::routes(google.clone(), service_cfg);
-    }
-    service_cfg.app_data(clients);
-}
-
-/// Allows the user to request another access token after it expired.
-/// This route checks the opaque refresh token against the database before granting (or not) the new access token.
-#[utoipa::path(
-    responses(
-        (status = StatusCode::OK, description = "Successfully refreshed", body = AuthenticationResponse),
-        (status = StatusCode::BAD_REQUEST, body = ApiError, examples(
-            ("Invalid path parameters" = (value = json!(path_error_example()))),
-            ("Invalid payload" = (value = json!(json_payload_error_example())))),
-            content_type = "application/json"
-        ),
-        (status = StatusCode::UNAUTHORIZED, description = "Credentials are invalid", body = AuthError, content_type = "application/json"),
-        (status = StatusCode::FORBIDDEN, description = "OAuth provider rejected the request", body = AuthError, content_type = "application/json"),
-    ),
-    tag="auth",
-    request_body(
-        description = "User id and refresh token",
-        content_type = "application/json",
-        content = RefreshTokenParams
-    )
-)]
-#[post("/auth/refresh")]
-#[tracing::instrument(skip_all)]
-pub async fn refresh_token(
-    db: web::Data<Database>,
-    json: web::Json<RefreshTokenParams>,
-    app_cfg: web::Data<AppConfiguration>,
-    clients: web::Data<OAuth2Clients>,
-) -> ActixResult<HttpResponse> {
-    let RefreshTokenParams {
-        user_id,
-        refresh_token,
-    } = json.into_inner();
-
-    // check in DB that there is a user that satisfies the following conditions:
-    // - user_id matches the one in the request
-    // - refresh_token matches the one in the request
-    // - refresh_token_exp is not expired
-    let result: Option<(String, Option<ClientType>, Option<String>)> =
-        database::open(db.clone(), move |conn| {
-            use crate::schema::users;
-            use diesel::dsl::now;
-            use diesel::prelude::*;
-
-            users::table
-                .filter(
-                    users::id.eq(user_id).and(
-                        users::refresh_token
-                            .eq(&refresh_token)
-                            .and(users::refresh_token_exp.ge(now)),
-                    ),
-                )
-                .select((users::email, users::auth_type, users::oauth_token))
-                .get_result(conn)
-                .optional()
-                .map_err(DatabaseError::from)
-        })
-        .await?;
-
-    let (email, client_type, oauth_token) =
-        result.ok_or_else(|| AuthErrorKind::InvalidCredentials.to_error())?;
-
-    check_user_authorized_oauth2(clients, client_type, oauth_token, &email).await?;
-
-    info!(%email, %user_id, lifetime_secs = app_cfg.access_token_lifetime.whole_seconds(), "refreshing access token of user");
-    let access_token_exp = OffsetDateTime::now_utc() + app_cfg.access_token_lifetime;
-    let access_token = AuthContext::encode_jwt(
-        user_id,
-        access_token_exp,
-        &app_cfg.app_name,
-        &app_cfg.jwt_secret,
-    )?;
-
-    // Regenerate a new refresh token and update the database
-    let user_info = database::open(db, move |conn| {
-        use crate::schema::users;
-        use diesel::prelude::*;
-
-        diesel::update(users::table.filter(users::id.eq(user_id)))
-            .set((
-                users::refresh_token.eq(&AuthContext::generate_refresh_token()),
-                users::refresh_token_exp
-                    .eq(OffsetDateTime::now_utc() + app_cfg.refresh_token_lifetime),
-            ))
-            .returning(PersistedUserInfo::as_returning())
-            .get_result(conn)
-            .map_err(DatabaseError::from)
-    })
-    .await?;
-
-    Ok(HttpResponse::Ok().json(AuthenticationResponse {
-        user: user_info,
-        access_token,
-        access_token_exp,
-    }))
-}
+pub use routes::configure;
 
 /// Sends a request to the OAuth2 provider to check if the user is still authorized.
 #[tracing::instrument(skip(clients, oauth_token))]
